Estava com conflito no banco de dados por ser long e no banco é int

+++ src/main/java/br/com/project/usuarios/Clientes.java
@@ -35,11 +35,11 @@ public class Clientes {
         this.gmail = gmail;
     }
 
-    public Long getIdCli() {
+    public int getIdCli() {
         return idCli;
     }
 
-    public void setIdCli(Long idCli) {
+    public void setIdCli(int idCli) {
         this.idCli = idCli;
     }
